requirement.txt
aiohttp>=3.8.0
Pillow>=9.0.0
numpy>=1.21.0